Fixed folders by tags endpoint not returning the expected response page object. Fixed not returning the full cover url and returning just the filename.
